common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Shared utilities for app/ and jobs/ (config, logging, GCS, BQ repository)"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8",
    "pydantic-settings[yaml]>=2.4",
    "google-cloud-bigquery>=3.25",
    "google-cloud-bigquery-storage>=2.27",
    "google-cloud-storage>=2.18",
    "google-cloud-logging>=3.11",
    "pandas>=2.2",
    "pyarrow>=16",
    "db-dtypes>=1.3",
    # E5Encoder lives in common/embeddings/e5_encoder.py and is used by both
    # app (lifespan when ENABLE_SEARCH=true) and jobs (embedding-job). Pulled
    # in here so neither package has to re-declare it.
    "sentence-transformers>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
